Выносит повторяющиеся элементы разметки в отдельный модуль
