+++ src/main.js
@@ -16,9 +16,9 @@ const cardsContainerElement = document.querySelector(`.board__tasks`);
 
 filterContainerElement.innerHTML = createFiltersTemplate(filters);
 
-cards.forEach((card) => {
-  const task = new Task(card);
-  const taskEdit = new TaskEdit(card);
+cards.forEach((card, index) => {
+  const task = new Task(card, index);
+  const taskEdit = new TaskEdit(card, index);
 
   let taskElement = task.render();
   let taskEditElement;
