+++ src/template/card-edit.js
@@ -1,60 +1,18 @@
-const createControlTemplate = () => {
-  const controls = [
-    {
-      name: `edit`,
-      isDisabled: ``
-    },
-    {
-      name: `archive`,
-      isDisabled: ``
-    },
-    {
-      name: `favourites`,
-      isDisabled: `card__btn--disabled`
-    }
-  ];
-
-  const controlsTemplate = controls
-    .map((control) => (
-      `<button
-        type="button"
-        class="card__btn card__btn--${control.name} ${(control.isDisabled) ? `disabled` : ``}"
-        >${control.name}</button>
-        `
-    ))
-    .join(``);
-
-  return (
-    `<div class="card__control">
-      ${controlsTemplate}
-    </div>`
-  );
-};
-
-const createDescriptionTemplate = (card) => (
-  `<div class="card__textarea-wrap">
-    <label>
-      <textarea
-        class="card__text"
-        placeholder="Start typing your text here..."
-        name="text">${card.title}</textarea>
-    </label>
-  </div>`
-);
-
-const createColorBarTemplate = (card) => (
-  `<div class="card__color-bar">
-    <svg class="card__color-bar-${card.colorSetting}" width="100%" height="10">
-      <use xlink:href="#wave"></use>
-    </svg>
-  </div>`
-);
+import {
+  createControlTemplate,
+  createColorBarTemplate,
+  createDescriptionTemplate,
+  createStatusBtnTemplate,
+  createHashtagTemplate,
+  createPictureTemplate,
+  createColorsTemplate
+} from './common-card-elements';
 
 const createDeadlineTemplate = (card) => (
   `<button class="card__date-deadline-toggle" type="button">
-    date: <span class="card__date-status">no</span>
+    date: <span class="card__date-status">${(card.dueDate.isEmpty) ? `no` : `yes`}</span>
   </button>
-  <fieldset class="card__date-deadline" disabled>
+  <fieldset class="card__date-deadline" ${card.dueDate.isEmpty ? `disabled` : ``}>
     <label class="card__input-deadline-wrap">
       <input
         class="card__date"
@@ -78,9 +36,9 @@ const createDeadlineTemplate = (card) => (
 
 const createRepeatingDaysTemplate = (card, index) => (
   `<button class="card__repeat-toggle" type="button">
-    repeat:<span class="card__repeat-status">no</span>
+    repeat:<span class="card__repeat-status">${card.checkRepeatingDays() ? `yes` : `no`}</span>
   </button>
-  <fieldset class="card__repeat-days" disabled>
+  <fieldset class="card__repeat-days" ${card.checkRepeatingDays() ? `` : `disabled`}>
     <div class="card__repeat-days-inner">
     ${[...card.repeatingDays.keys()].map((day) => (
     `<input
@@ -98,10 +56,10 @@ const createRepeatingDaysTemplate = (card, index) => (
   </fieldset>`
 );
 
-const createDatesTemplate = (card) => (`
+const createDatesTemplate = (card, index) => (`
   <div class="card__dates">
     ${createDeadlineTemplate(card)}
-    ${createRepeatingDaysTemplate(card)}
+    ${createRepeatingDaysTemplate(card, index)}
   </div>`
 );
 
@@ -112,91 +70,11 @@ const createDetailsTemplate = (card, index) => (`
   </div>
 `);
 
-const createPictureTemplate = (card) => (`
-  <label class="card__img-wrap card__img-wrap--empty">
-    <input
-      type="file"
-      class="card__img-input visually-hidden"
-      name="img"
-    />
-    <img
-      src="${card.picture}"
-      alt="task picture"
-      class="card__img"
-    />
-  </label>
-`);
-
-const createColorsTemplate = (card, index) => (`
-  <div class="card__colors-inner">
-    <h3 class="card__colors-title">Color</h3>
-    <div class="card__colors-wrap">
-      <input
-            type="radio"
-            id="color-${card.color}-${index}"
-            class="card__color-input card__color-input--${card.color} visually-hidden"
-            name="color"
-            value="${card.color}"
-          />
-          <label
-            for="color-${card.color}-${index}"
-            class="card__color card__color--${card.color}"
-            >${card.color}
-          </label>
-    </div>
-  </div>
-`);
-
 const createSettingsTemplate = (card, index) => (`
   <div class="card__settings">
     ${createDetailsTemplate(card, index)}
     ${createPictureTemplate(card)}
     ${createColorsTemplate(card, index)}
-  </div>
-`);
-
-const createHashtagButtonTemplate = (hashtag) => `
-  <button type="button" class="card__hashtag-name">
-    #${hashtag}
-  </button>
-  <button type="button" class="card__hashtag-delete">
-    delete
-  </button>
-`;
-
-const createHashtagTemplate = (card) => (
-  `<div class="card__hashtag">
-    <div class="card__hashtag-list">
-      ${[...card.tags.keys()]
-        .map((tag) => (
-          `<span class="card__hashtag-inner">
-          <input
-            type="hidden"
-            name="hashtag"
-            value="${tag}"
-            class="card__hashtag-hidden-input"
-          />
-          ${createHashtagButtonTemplate(tag)}
-        </span>`
-        ))
-        .join(``)
-  }
-    </div>
-    <label>
-      <input
-        type="text"
-        class="card__hashtag-input"
-        name="hashtag-input"
-        placeholder="Type new hashtag here"
-      />
-    </label>
-  </div>`
-);
-
-const createStatusBtnTemplate = () => (`
-  <div class="card__status-btns">
-    <button class="card__save" type="submit">save</button>
-    <button class="card__delete" type="button">delete</button>
   </div>
 `);
